[FE] utils, service, hooks 폴더 정리 (jsdoc 활용) (#183)
* Build: jsdoc 설치

* Docs: useForm에 jsdoc 주석 작성

* Docs: login.valdation jsdoc 주석 작성

* Docs: singup.valdation jsdoc 주석 작성

* Rename: login.validation, signup.valiation 통합

-> auth.validation

* Rename: auth.parser, auth.validation. auth.js 폴더로 묶음

* Rename: editUserValidation, essentialValidation 통합

useValidation

* Docs: user.validation jsdoc 주석 작성

* Rename: swagger 최신화에 따른 userDetailTypeSchema변경

* Docs: jsdoc용 유저 type

* Rename: user.parser 폴더 이동 -> service/user로 이동

* Fix: MarkdownEditor viewer용 props 설정

name, onChange, placeholder -> 필수조건에서 제거

* Fix: TechStackOptions 타입변경

object -> array

* Fix: 유저필수정보 및 유저디테일 mock 데이터 swagger최신화

* Feat: 유저 정보 입력 스키마 추가

* Docs: userParser에 jsdoc 주석 작성

* Rename: user.validtaion 함수 jsdocs 수정

구체적인 네이밍으로 변경

* Docs: team.validation jsdoc 주석 작성

* Docs: usertype 주석 수정

* Docs: 팀 타입 Jsdoc 주석 작성

* Docs: team parser jsdoc 작성

* Rename: src/serivce/team으로 묶음

team.validation
team.parser

* Docs: skillStack 타입 jsdoc 주석 짝성

* Docs: skillStack parser jsdoc 주석 짝성

* Rename: skillStack.parser 이동 (src/serivce/etc)

* Docs: useAxios jsdoc 주석 작성

* Docs: useCheckNicknameDuplicate jsdoc 작성

* Docs: useIntersect를 jsdoc 작성

* Docs: useDropdown jsdoc 주석 작성

* Docs: useFileUploader jsdoc 주석 작성

* Docs: useSkillCarousel jsdoc 주석 작성

* Docs: useUserInfo jsdoc 주석 작성

* Docs: utils함수들 jsdoc주석 작성

* Chore: setDefaultProfileImage 함수제거
