+++ webapp/src/components/ComentContainer/CommentElement/Nested.CommentElement.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { setDefaultProfileImage } from 'utils';
 import { getUserInfo } from 'service/auth';
 import { useCommentsAction, useCommentsState } from 'contexts/Comment/Comment.Provider';
 import { commentInfoType } from 'types/comment.type';
@@ -38,7 +37,7 @@ export default function NestedCommentElement({ commentId, commentInfo }) {
     <S.NestedCommentBox>
       <S.NormalCommentBox style={{ display: 'flex' }}>
         <S.UserInfo>
-          <img src={setDefaultProfileImage(img)} alt="profile" />
+          <img src={img} alt="profile" />
           <h3>{commenWriter}</h3>
         </S.UserInfo>
         {!isEditTargetComment && (
