+++ webapp/src/contexts/EssentialForm/useEssentialForm.js
@@ -7,11 +7,12 @@ import useForm from 'hooks/useForm';
 import authApi from 'api/auth.api';
 
 import { ROUTE } from 'constant/route.constant';
-import essentialValidation from 'service/essentialForm.validation';
-import { essentialInfoParser } from 'service/user.parser';
+
+import { essentialInfoParser } from 'service/user/user.parser';
 import useUserInfo from 'hooks/useUserInfo';
 import useFileUploader from 'hooks/useFileUploader';
 import userApi from 'api/user.api';
+import { essentialValidation } from 'service/user/user.validation';
 
 const initialValues = {
   nickname: '',
