+++ webapp/src/hooks/useFileUploader.js
@@ -7,6 +7,21 @@ import { useNavigate } from 'react-router-dom';
 import { ROUTE } from 'constant/route.constant';
 import useUserInfo from './useUserInfo';
 
+/**
+ * useFileUploader를 사용하는 곳에서 사용할 method 및 state
+ * @typedef {Object} useFileUploaderReturns
+ * @property {string} imageFile isLoading, responseData, error로 구성
+ * @property {(Event) => void} onChangeFile 새로운 axios config와 함께 api호출을 실행하는 함수
+ * @property {string} s3ImageId 기존 axios config로 같은 api호출을 실행하는 함수
+ * @property {Object} s3ImageObj id: s3ImageId, path: /image/:s3ImageId
+ * @property {(submitImageFile: File) => Promise<{id: string, path: string} | null>} uploadFileOnS3 s3에 이미지 업로드 요청하는 함수
+ * @property {() => Promise<void>} deleteFileOnS3  s3에 올라간 이미지를 제거 요청하는 함수
+ */
+
+/**
+ * 파일 업로드 (s3)를 위한 cutom hooks
+ * @returns {useFileUploaderReturns} useFileUploader를 사용하는 곳에서 사용할 method 및 state
+ */
 const useFileUploader = () => {
   const notifyDispatch = useToastNotificationAction();
   const navigate = useNavigate();
@@ -15,11 +30,20 @@ const useFileUploader = () => {
   const [s3ImageObj, setS3ImageObj] = useState(null);
   const [imageFile, setImageFile] = useState(null);
 
+  /**
+   * input[type="file"]을 변경을 위한 change handler
+   * @param {Event}
+   */
   const onChangeFile = (event) => {
     const targetImageFile = event.target.files[0];
     setImageFile(targetImageFile);
   };
 
+  /**
+   * s3에 이미지 업로드 요청하는 함수
+   * @param {File} 새롭게 제출할 이미지 파일
+   * @returns {Promise<void>}
+   */
   const uploadFileOnS3 = async (submitImageFile) => {
     // 둘다 없을 때 실행하지 않는다.
     if (!imageFile && !submitImageFile) {
@@ -48,6 +72,11 @@ const useFileUploader = () => {
     }
   };
 
+  // FIXME: 아직 구현 안 됨. 백엔드랑 얘기하고 수정해야함.
+  /**
+   * s3에 올라간 이미지를 제거 요청하는 함수
+   * @returns {Promise<void>}
+   */
   const deleteFileOnS3 = async () => {
     try {
       const response = await etcApi.deleteImage();
