+++ webapp/src/mocks/userHandler.mock/essentialInfo.js
@@ -4,7 +4,7 @@ export const mockEssentialInfo = {
   portfolio: 'www.포트폴리오링크.com',
   slogan: 'string',
   content: '임시 자기 소개',
-  img: 'string',
+  image: 'string',
   hopeSession: '1개월 이하',
   job: '학생',
   status: false,
