+++ webapp/src/pages/NewTeamPost/NewTeamPostDetail.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { newTeamPostParser } from 'service/team.parser';
-import { newTeamPostValidation } from 'service/team.validation';
+import { newTeamPostParser } from 'service/team/team.parser';
+import { newTeamPostValidation } from 'service/team/team.validation';
 import { useToastNotificationAction } from 'contexts/ToastNotification';
 import { notifyNewMessage } from 'contexts/ToastNotification/action';
 import { TOAST_TYPE } from 'contexts/ToastNotification/type';
