+++ webapp/src/service/team/team.validation.js
@@ -1,0 +1,41 @@
+/**
+ * 팀 공고글 생성: Inputs Object - input들의 key, value로 이뤄진 객체
+ * @typedef {Object} newTeamPostValidateObj
+ * @property {string} name  nickname
+ * @property {Array} techSkills  techSkills
+ * @property {string} slogan slogan
+ */
+
+/**
+ * 팀 공고글 생성: Erros Object - input들의 key, value로 이뤄진 에러 객체
+ * @typedef {Object} newTeamPostValidateErrors
+ * @property {string} nickname
+ * @property {string} techSkills
+ * @property {string} slogan
+ */
+
+/**
+ * 팀 공고글 생성 vaidation
+ * @param {newTeamPostValidateObj} newTeamPostValidateObj input들의 key, value로 이뤄진 객체
+ * @returns {newTeamPostValidateErrors} input들의 key, value로 이뤄진 에러 객체
+ */
+export const newTeamPostValidation = ({ name, techSkills, slogan }) => {
+  const validateErrors = {
+    name: '',
+    techSkills: '',
+    slogan: '',
+  };
+
+  if (!name) {
+    validateErrors.name = '팀이름이 입력되지 않았습니다. ';
+  }
+
+  if (techSkills.length === 0) {
+    validateErrors.techSkills = '최소 한 개 이상의 언어를 선택해주세요.';
+  }
+  if (!slogan) {
+    validateErrors.slogan = '슬로건이 입력되지 않았습니다. ';
+  }
+
+  return validateErrors;
+};
