+++ webapp/src/types/user.type.js
@@ -7,6 +7,9 @@ export const loggedUserType = shape({
   profileImg: PropTypes.string,
 });
 
+/**
+ * @type {UserSchema}
+ */
 const userSchema = {
   id: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
@@ -15,38 +18,36 @@ const userSchema = {
   status: PropTypes.bool.isRequired,
 };
 
+/**
+ * @type {UserCardSchema}
+ */
 const userCardTypeSchema = {
   ...userSchema,
   commentCnt: PropTypes.number.isRequired,
   likeCnt: PropTypes.number.isRequired,
 };
 
+/**
+ * @type {UserDetailSchema}
+ */
 const userDetailTypeSchema = {
-  ...userSchema,
-  job: PropTypes.string.isRequired,
-  img: PropTypes.string.isRequired,
-  hopeSession: PropTypes.string.isRequired,
-  description: PropTypes.string.isRequired,
+  id: PropTypes.number.isRequired,
+  skills: PropTypes.arrayOf(rawResponseTechStackType).isRequired,
+  slogan: PropTypes.string.isRequired,
+  status: PropTypes.bool.isRequired,
   commentCnt: PropTypes.number.isRequired,
   content: PropTypes.string.isRequired,
   email: PropTypes.string.isRequired,
+  hopeSession: PropTypes.string.isRequired,
+  job: PropTypes.string.isRequired,
   likeCnt: PropTypes.number.isRequired,
-  oauthId: PropTypes.string.isRequired,
   portfolio: PropTypes.string.isRequired,
-};
-
-const userEditTypeSchema = {
-  ...userSchema,
-  job: PropTypes.string,
-  img: PropTypes.string,
-  hopeSession: PropTypes.string,
-  description: PropTypes.string,
-  content: PropTypes.string,
-  portfolio: PropTypes.string,
-  email: PropTypes.string.isRequired,
-  commentCnt: PropTypes.number.isRequired,
-  likeCnt: PropTypes.number.isRequired,
-  oauthId: PropTypes.string.isRequired,
+  readCnt: PropTypes.number.isRequired,
+  userInfo: PropTypes.shape({
+    id: PropTypes.number.isRequired,
+    image: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+  }),
 };
 
 export const userType = shape(userSchema);
@@ -54,5 +55,3 @@ export const userType = shape(userSchema);
 export const userCardType = shape(userCardTypeSchema);
 
 export const userDetailType = shape(userDetailTypeSchema);
-
-export const userEditType = shape(userEditTypeSchema);
